% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ggplots.R
\name{asdv_internalFit}
\alias{asdv_internalFit}
\title{Linear model fitting of channel A and B time series data}
\usage{
asdv_internalFit(
  pat = NULL,
  showPlot = TRUE,
  whichPlot = "lm",
  size = 1,
  a_color = "red",
  b_color = "blue",
  alpha = 0.25,
  lr_shape = 15,
  lr_color = "black",
  lr_lwd = 1.5,
  lr_lcolor = "tomato",
  lr_lalpha = 0.45,
  ts_shape = 1,
  xylim = NULL,
  tz = NULL
)
}
\arguments{
\item{pat}{PurpleAir Timeseries \emph{pat} object.}

\item{showPlot}{Logical specifying whether to generate a model fit plot.}

\item{whichPlot}{Which plot to show.}

\item{size}{Size of points.}

\item{a_color}{Color of time series channel A points.}

\item{b_color}{Color of time series channel B points.}

\item{alpha}{Opacity of points.}

\item{lr_shape}{Symbol to use for linear regression points.}

\item{lr_color}{Color of linear regression points.}

\item{lr_lwd}{Width of linear regression line.}

\item{lr_lcolor}{Color of linear regression line.}

\item{lr_lalpha}{Opacity of linear regression line.}

\item{ts_shape}{Symbol to use for time series points.}

\item{xylim}{Vector of (lo,hi) limits used as limits on the correlation plot
axes -- useful for zooming in.}

\item{tz}{a timezone.}
}
\value{
A linear model, fitting the `pat` B channel readings to A channel
readings.
}
\description{
Uses a linear model to fit data from channel B to data from
channel A.

A diagnostic plot is produced if `showPlot = TRUE`.
}
\examples{
\donttest{
asdv_internalFit(pat = example_pat)
}
}
