% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotlys.R
\name{channelPlotly}
\alias{channelPlotly}
\title{Channel Graph Plotly}
\usage{
channelPlotly(pat, channel = "ab")
}
\arguments{
\item{pat}{a pa_timeseries object}

\item{channel}{pm channel}
}
\value{
a plotly
}
\description{
Channel Graph Plotly
}
