% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{com2id}
\alias{com2id}
\title{Community name to 4 digit code}
\usage{
com2id(X)
}
\arguments{
\item{X}{a friendly community string.}
}
\value{
an unfriendly community string.
}
\description{
Community name to 4 digit code
}
