% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_leaflet.R
\name{comparisonLeaflet}
\alias{comparisonLeaflet}
\title{Comparison Leaflet}
\usage{
comparisonLeaflet(sensor, pwfsl)
}
\arguments{
\item{sensor}{An airsensor object.}

\item{pwfsl}{A ws_monitor object.}
}
\value{
a leaflet map
}
\description{
Map an airsensor and ws_monitor object.
}
