% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_noaa}
\alias{get_noaa}
\title{Get NOAA Data From ISD
 
 A homebrew solution to the acquiring NOAA data from their database, following
 the worldmet importNOAA function.}
\usage{
get_noaa(sensor, sd, ed)
}
\arguments{
\item{sensor}{A univariate airsensor object}

\item{sd}{A startdate.}

\item{ed}{A enddate.}
}
\value{
a data.frame
}
\description{
Get NOAA Data From ISD
 
 A homebrew solution to the acquiring NOAA data from their database, following
 the worldmet importNOAA function.
}
