% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_pas}
\alias{get_pas}
\title{Get PurpleAir Synoptic Data}
\usage{
get_pas(datestamp = NULL)
}
\arguments{
\item{datestamp}{a datestamp (YYYYmmdd)}
}
\value{
a pa_synoptic object
}
\description{
Get PurpleAir Synoptic Data
}
