% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_pat}
\alias{get_pat}
\title{Get PurpleAir Timeseries Data}
\usage{
get_pat(pas, label, sd, ed, pat = NULL)
}
\arguments{
\item{pas}{a pas object.}

\item{label}{a sensor label}

\item{sd}{a startdate (YYYYmmdd)}

\item{ed}{a enddate (YYYYmmdd)}

\item{pat}{an optional pat to attempt filtering on.}
}
\value{
a pa_timeseries object
}
\description{
Get PurpleAir Timeseries Data
}
