% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_pwfsl}
\alias{get_pwfsl}
\title{Get PWFSL Monitor Data}
\usage{
get_pwfsl(sd, ed, id)
}
\arguments{
\item{sd}{a startdate (YYYYmmdd)}

\item{ed}{an enddate (YYYYmmdd)}

\item{id}{a valid PWFSL Monitor ID}
}
\value{
a ws_monitor object
}
\description{
Get PWFSL Monitor Data
}
