% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_sensor}
\alias{get_sensor}
\title{Get A Sensor from Sensors}
\usage{
get_sensor(sensors, ...)
}
\arguments{
\item{sensors}{an airsensor object}

\item{label}{a label to filter the airsensor object by}

\item{sd}{a startdate (YYYYmmdd)}

\item{ed}{an enddate (YYYYmmdd)}
}
\value{
a univariate airsensor object
}
\description{
Get A Sensor from Sensors
}
