% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load.R
\name{get_sensors}
\alias{get_sensors}
\title{Get Sensors for date range}
\usage{
get_sensors(sd, ed, sensors = NULL)
}
\arguments{
\item{sd}{a startdate (YYYYmmdd)}

\item{ed}{an enddate (YYYYmmdd)}
}
\value{
an airsensor object
}
\description{
Get Sensors for date range
}
