% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotlys.R
\name{humidityPlotly}
\alias{humidityPlotly}
\title{Humidity Graph Plotly}
\usage{
humidityPlotly(pat)
}
\arguments{
\item{pat}{a pa_timeseries object}
}
\value{
a plotly
}
\description{
Humidity Graph Plotly
}
