% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadNOAA.R
\name{loadNOAA}
\alias{loadNOAA}
\title{Load NOAA data from Sensor Data}
\usage{
loadNOAA(sensor)
}
\arguments{
\item{sensor}{A sensor obj.}

\item{startdate}{Startdate ISO}

\item{enddate}{Enddate ISO}

\item{tz}{timezone}
}
\value{
NOAA data frame
}
\description{
Load NOAA data from Sensor Data
}
