% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stateman.R
\name{magicUpdate}
\alias{magicUpdate}
\title{A "Magic" Auto Updater}
\usage{
magicUpdate(usr)
}
\arguments{
\item{usr}{}
}
\value{

}
\description{
A tool  intended to minimize local computational cost that handles the current 
client object state to best determine what to update.
}
