% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{makeWaitress}
\alias{makeWaitress}
\title{A Waitress notification wrapper}
\usage{
makeWaitress(expr, msg)
}
\arguments{
\item{expr}{an expression to make loading notification for}

\item{msg}{a message to show the user.}
}
\description{
A Waitress notification wrapper
}
