% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{noaaTable}
\alias{noaaTable}
\title{NOAA table constructor}
\usage{
noaaTable(noaa)
}
\arguments{
\item{noaa}{a noaa data frame generated from get_noaa().}
}
\value{
a data.frame
}
\description{
NOAA table constructor
}
