% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R, R/User_OLLD.R
\name{reactiveTrigger}
\alias{reactiveTrigger}
\title{R6 reactive trigger: 
https://gist.github.com/bborgesr/3350051727550cfa798cb4c9677adcd4}
\usage{
reactiveTrigger()

reactiveTrigger()
}
\value{



}
\description{
R6 reactive trigger: 
https://gist.github.com/bborgesr/3350051727550cfa798cb4c9677adcd4

R6 reactive trigger: https://gist.github.com/bborgesr/3350051727550cfa798cb4c9677adcd4
}
