% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{sensorMonitorCompTable}
\alias{sensorMonitorCompTable}
\title{Monitor Sensor Comparison Table}
\usage{
sensorMonitorCompTable(pat)
}
\arguments{
\item{pat}{pa timeseries obh}
}
\value{

}
\description{
Monitor Sensor Comparison Table
}
