% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logs.R
\name{setupSessionLogs}
\alias{setupSessionLogs}
\title{Setup Logging for Shiny Sessions}
\usage{
setupSessionLogs(session)
}
\arguments{
\item{session}{a shiny session object.}
}
\description{
Setup Logging for Shiny Sessions
}
