% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ggplots.R
\name{stat_meanByHour}
\alias{stat_meanByHour}
\title{Add hourly averages to a plot}
\usage{
stat_meanByHour(
  mapping = NULL,
  data = NULL,
  input = NULL,
  output = "y",
  geom = "bar",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If
specified and \code{inherit.aes = TRUE} (the default), it is combined with
the default mapping at the top level of the plot. You must supply
\code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
if \code{NULL}, the default, the data is inherited from the plot data. A
\code{data.frame} or other object, will override the plot data. A
\code{function} will be called with a single argument, the plot data. The
return value must be a \code{data.frame}, and will be used as the layer
data.}

\item{input}{The value to find the mean of. If \code{NULL}, the default
\code{y} value will be used.}

\item{output}{"AQIColors", "mv4Colors", "scaqmd", "y"}

\item{geom}{The geometic object to display the data}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{na.rm}{remove NA values from data}

\item{show.legend}{logical indicating whether this layer should be included
in legends.}

\item{inherit.aes}{if \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and the aesthetics and shouldn't inherit behaviour from
the default plot specificatino, eg \code{borders()}.}

\item{...}{additional arguments passed on to \code{layer()}, such as
aesthetics.}
}
\description{
This function calculates the mean y-value for each x-value. Should be used
only when \code{x} is discrete. The resulting mean can be mapped to any
aesthetic, specified with the \code{output} parameter.
}
