% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotlys.R
\name{temperaturePlotly}
\alias{temperaturePlotly}
\title{Temperature Graph Plotly}
\usage{
temperaturePlotly(pat)
}
\arguments{
\item{pat}{a pa_timeseries object}
}
\value{
a plotly
}
\description{
Temperature Graph Plotly
}
