% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ggplots.R
\name{theme_pwfsl}
\alias{theme_pwfsl}
\title{Theme for PWFSL plots}
\usage{
theme_pwfsl(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A \emph{ggplot} theme.
}
\description{
Applies the package standard theme to a \emph{ggplot} plot object.
}
