% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{User}
\alias{User}
\title{User R6 Class}
\description{
Create a user to act as a data and state storage object for AirSensor 
DataViewer Shiny Application.
}
\details{
Create an object that is responsible for housing the core state and data components of a 
Shiny applications user session. Each user is dependent on a Shiny session object. 
The methods and data of the object can be shared across modules without the requirment 
to create reactive-values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-updatePas}{\code{User$updatePas()}}
\item \href{#method-updateSensors}{\code{User$updateSensors()}}
\item \href{#method-updatePat}{\code{User$updatePat()}}
\item \href{#method-updateSensor}{\code{User$updateSensor()}}
\item \href{#method-updatePwfsl}{\code{User$updatePwfsl()}}
\item \href{#method-updateLatest}{\code{User$updateLatest()}}
\item \href{#method-updateAnnual}{\code{User$updateAnnual()}}
\item \href{#method-updateNoaa}{\code{User$updateNoaa()}}
\item \href{#method-setTz}{\code{User$setTz()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePas"></a>}}
\if{latex}{\out{\hypertarget{method-updatePas}{}}}
\subsection{Method \code{updatePas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePas(date = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateSensors"></a>}}
\if{latex}{\out{\hypertarget{method-updateSensors}{}}}
\subsection{Method \code{updateSensors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateSensors(sd, ed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePat"></a>}}
\if{latex}{\out{\hypertarget{method-updatePat}{}}}
\subsection{Method \code{updatePat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePat(label, sd, ed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateSensor"></a>}}
\if{latex}{\out{\hypertarget{method-updateSensor}{}}}
\subsection{Method \code{updateSensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateSensor(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePwfsl"></a>}}
\if{latex}{\out{\hypertarget{method-updatePwfsl}{}}}
\subsection{Method \code{updatePwfsl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePwfsl(label, sd, ed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateLatest"></a>}}
\if{latex}{\out{\hypertarget{method-updateLatest}{}}}
\subsection{Method \code{updateLatest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateLatest(label, tz = "UTC")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateAnnual"></a>}}
\if{latex}{\out{\hypertarget{method-updateAnnual}{}}}
\subsection{Method \code{updateAnnual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateAnnual(date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateNoaa"></a>}}
\if{latex}{\out{\hypertarget{method-updateNoaa}{}}}
\subsection{Method \code{updateNoaa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateNoaa(sd, ed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTz"></a>}}
\if{latex}{\out{\hypertarget{method-setTz}{}}}
\subsection{Method \code{setTz()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$setTz(timezone)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
