% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{withLoader}
\alias{withLoader}
\title{A CSS loader feedback wrapper function}
\usage{
withLoader(el)
}
\arguments{
\item{el}{an element to append to.}
}
\description{
A CSS loader feedback wrapper function
}
