% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{cacheKey}
\alias{cacheKey}
\title{Create a hash-cache key for caching}
\usage{
cacheKey(...)
}
\arguments{
\item{x}{an object, expression, anything.}
}
\value{
a hash
}
\description{
Create a hash-cache key for caching
}
