% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ggplots.R
\name{lmSensorMonitor}
\alias{lmSensorMonitor}
\title{Sensor Monitor Correlation Plot}
\usage{
lmSensorMonitor(sensor, pwfsl)
}
\arguments{
\item{sensor}{A airsensor object}

\item{pwfsl}{A ws_monitor object}
}
\value{

}
\description{
Sensor Monitor Correlation Plot
}
