% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{User}
\alias{User}
\title{User R6 Class}
\description{
Create a user to act as a data and state storage object for AirSensor 
DataViewer Shiny Application.
}
\details{
Create an object that is responsible for housing the core state and data components of a 
Shiny applications user session. Each user is dependent on a Shiny session object. 
The methods and data of the object can be shared across modules without the requirment 
to create reactive-values.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{User$new()}}
\item \href{#method-updatePas}{\code{User$updatePas()}}
\item \href{#method-updateSensors}{\code{User$updateSensors()}}
\item \href{#method-updatePat}{\code{User$updatePat()}}
\item \href{#method-updateSensor}{\code{User$updateSensor()}}
\item \href{#method-updatePwfsl}{\code{User$updatePwfsl()}}
\item \href{#method-updateLatest}{\code{User$updateLatest()}}
\item \href{#method-updateAnnual}{\code{User$updateAnnual()}}
\item \href{#method-updateNoaa}{\code{User$updateNoaa()}}
\item \href{#method-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(session, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}

\item{\code{...}}{Additional options}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\code{User$pas}: Return a future pa-synoptic object.


\code{User$sensors}: Return a future airsensor object.


\code{User$pat}: Return a future pa_timeseries object.


\code{User$sensor}: Return a future filtered airsensor object.


\code{User$pwfsl}: Return a future ws_monitor object from the PWFSLSmoke Package.


\code{User$latest}: Return a future latest pa-timeseries object (from \code{pat_createNew()}).


\code{User$annual}: Return an annual future airsensor object.


\code{User$noaa}: Return a future dataframe generated from NOAA Database.


\code{User$selected$...}: Access Shiny Input reactive values.


\code{User$url}: Access the current page URL for bookmarking state.


\code{User$token}: Access the current session token ID.


\code{User$url}: Access the {waiter} object for controling visual loading state.


Create a user.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePas"></a>}}
\if{latex}{\out{\hypertarget{method-updatePas}{}}}
\subsection{Method \code{updatePas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePas(date = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{YYYYmmdd.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users pa_synoptic data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateSensors"></a>}}
\if{latex}{\out{\hypertarget{method-updateSensors}{}}}
\subsection{Method \code{updateSensors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateSensors(sd, ed, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sd}}{startdate YYYYmmdd.}

\item{\code{ed}}{enddate YYYYmmdd.}

\item{\code{...}}{additional params.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users airsensor object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePat"></a>}}
\if{latex}{\out{\hypertarget{method-updatePat}{}}}
\subsection{Method \code{updatePat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePat(label, sd, ed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sd}}{startdate YYYYmmdd.}

\item{\code{ed}}{enddate YYYYmmdd.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users pa_timeseries object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateSensor"></a>}}
\if{latex}{\out{\hypertarget{method-updateSensor}{}}}
\subsection{Method \code{updateSensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateSensor(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{the sensor label to filter the airsensor object by.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users filtered airsensor object. 
Commonly used when a user selects a different sensor.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updatePwfsl"></a>}}
\if{latex}{\out{\hypertarget{method-updatePwfsl}{}}}
\subsection{Method \code{updatePwfsl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updatePwfsl(label, sd, ed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{?}

\item{\code{sd}}{startdate YYYYmmdd.}

\item{\code{ed}}{enddate YYYYmmdd.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users ws_monitor object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateLatest"></a>}}
\if{latex}{\out{\hypertarget{method-updateLatest}{}}}
\subsection{Method \code{updateLatest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateLatest(label, tz = getOption("asdv.timezone"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{a label of the sensor.}

\item{\code{tz}}{a specified timezone. (default: UTC)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users latest pa_timeseries object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateAnnual"></a>}}
\if{latex}{\out{\hypertarget{method-updateAnnual}{}}}
\subsection{Method \code{updateAnnual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateAnnual(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{a date to parse year from.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users annual airsensor object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateNoaa"></a>}}
\if{latex}{\out{\hypertarget{method-updateNoaa}{}}}
\subsection{Method \code{updateNoaa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$updateNoaa(sd, ed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sd}}{startdate YYYYmmdd.}

\item{\code{ed}}{enddate YYYYmmdd.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Update the users NOAA data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
